chore: New Google Client ID for iOS library
Update 'OSSocialLoginsLib.xcframework' to the latest version. For Google Sign-in, this requires a new property for the 'Info.plist' file: 'GIDClientID'. This is set, if configured, through the 'iOSCopyPreference' hook, based on the Google URL Scheme value.
Do some refactoring to the 'iOSCopyPreferences' hook and 'plugin.xml', removing redundant code and unneeded placeholders.

References: https://outsystemsrd.atlassian.net/browse/RMET-2800